requirements.txt
pyquaternion
pyyaml
pexpect
dm_control==1.0.14
mujoco==3.2.3
opencv-python
matplotlib
einops
packaging
h5py
jupyter
ipython
ftfy
transformers==4.49.0
sentencepiece==0.2.0
accelerate
diffusers
timm
wandb
pytransform3d
git+https://github.com/facebookresearch/pytorch3d.git@stable
git+https://github.com/RogerQi/maskclip_onnx
pytest
opentv
plotly
pin-pink
qpsolvers
